+++ src/lsp-server/schema.ts
@@ -24,6 +24,15 @@ export async function initializeSchema(schemaConfig: {
 
 function setSchema(schemaConfig: { versionName: string; customPath: string }) {
   let schemaPath: string = schemaConfig.customPath;
+  if (schemaPath !== "" && schemaConfig.versionName !== "Custom") {
+    console.warn(
+      "Custom schema path provided, but version is not set to Custom.  Ignoring custom path."
+    );
+  }
+  if (!fs.existsSync(schemaPath)) {
+    console.error(`Schema file not found at path: ${schemaPath}`);
+    schemaPath = "";
+  }
   if (schemaPath === "") {
     switch (schemaConfig.versionName) {
       case "Stable":
@@ -34,7 +43,7 @@ function setSchema(schemaConfig: { versionName: string; customPath: string }) {
         break;
       case "Custom":
         console.log(
-          "Selected custom schema, but no path provided.  Setting to default."
+          "Selected custom schema, but no valid path provided.  Setting to default."
         );
         schemaPath = path.join(schemaDir, "pretext.rng");
         break;